+++ diet_student/lib/components/pie_chart.dart
@@ -11,9 +11,9 @@ class Task {
 }
 
 class PieChart extends StatefulWidget {
-  final int proteins;
-  final int lipids;
-  final int glucides;
+  final double proteins;
+  final double lipids;
+  final double glucides;
 
   @override
   _PieChartState createState() => _PieChartState(proteins, lipids, glucides);
@@ -23,9 +23,9 @@ class PieChart extends StatefulWidget {
 
 class _PieChartState extends State<PieChart>
     with SingleTickerProviderStateMixin {
-  final int proteins;
-  final int lipids;
-  final int glucides;
+  final double proteins;
+  final double lipids;
+  final double glucides;
   // chart
   List<charts.Series<Task, String>> _seriesPieData;
 
@@ -38,10 +38,10 @@ class _PieChartState extends State<PieChart>
     //chart
     _seriesPieData = List<charts.Series<Task, String>>();
     var piedata = [
-      new Task('Protéines', proteins.toDouble(), Color(0xff3366cc)),
+      new Task('Protéines', proteins, Color(0xff3366cc)),
       //new Task('Lipides', 8.3, Color(0xff990099)),
-      new Task('Glucides', glucides.toDouble(), Color(0xff109618)),
-      new Task('Lipides', lipids.toDouble(), Color(0xfffdbe19)),
+      new Task('Glucides', glucides, Color(0xff109618)),
+      new Task('Lipides', lipids, Color(0xfffdbe19)),
       //new Task('Sel', 19.2, Color(0xffff9900)),
       //new Task('Other', 10.3, Color(0xffdc3912)),
     ];
